#6581: lazy load page script
